+++ core/src/v2/schema.rs
@@ -1,6 +1,8 @@
 //! Traits used for code and spec generation.
 
-use super::models::{DataType, DataTypeFormat, DefaultSchemaRaw, Either, Operation, SchemaRepr};
+use super::models::{
+    DataType, DataTypeFormat, DefaultOperationRaw, DefaultSchemaRaw, Either, Resolvable,
+};
 
 use std::collections::{BTreeMap, BTreeSet};
 
@@ -24,22 +26,22 @@ pub trait Schema: Sized {
     fn format(&self) -> Option<&DataTypeFormat>;
 
     /// Schema for array definitions, if any (`items` field).
-    fn items(&self) -> Option<&Either<SchemaRepr<Self>, Vec<SchemaRepr<Self>>>>;
+    fn items(&self) -> Option<&Either<Resolvable<Self>, Vec<Resolvable<Self>>>>;
 
     /// Mutable access to the `items` field, if it exists.
-    fn items_mut(&mut self) -> Option<&mut Either<SchemaRepr<Self>, Vec<SchemaRepr<Self>>>>;
+    fn items_mut(&mut self) -> Option<&mut Either<Resolvable<Self>, Vec<Resolvable<Self>>>>;
 
     /// Value schema for maps (`additional_properties` field).
-    fn additional_properties(&self) -> Option<&Either<bool, SchemaRepr<Self>>>;
+    fn additional_properties(&self) -> Option<&Either<bool, Resolvable<Self>>>;
 
     /// Mutable access to `additional_properties` field, if it's a map.
-    fn additional_properties_mut(&mut self) -> Option<&mut Either<bool, SchemaRepr<Self>>>;
+    fn additional_properties_mut(&mut self) -> Option<&mut Either<bool, Resolvable<Self>>>;
 
     /// Map of names and schema for properties, if it's an object (`properties` field)
-    fn properties(&self) -> Option<&BTreeMap<String, SchemaRepr<Self>>>;
+    fn properties(&self) -> Option<&BTreeMap<String, Resolvable<Self>>>;
 
     /// Mutable access to `properties` field.
-    fn properties_mut(&mut self) -> Option<&mut BTreeMap<String, SchemaRepr<Self>>>;
+    fn properties_mut(&mut self) -> Option<&mut BTreeMap<String, Resolvable<Self>>>;
 
     /// Returns the required properties (if any) for this object.
     fn required_properties(&self) -> Option<&BTreeSet<String>>;
@@ -328,7 +330,7 @@ impl_schema_map!(BTreeMap<K, V>);
 /// but *can* be used for constraining stuff in framework-related impls.
 pub trait Apiv2Operation<T, R> {
     /// Returns the definition for this operation.
-    fn operation() -> Operation<DefaultSchemaRaw>;
+    fn operation() -> DefaultOperationRaw;
 
     /// Returns the definitions used by this operation.
     fn definitions() -> BTreeMap<String, DefaultSchemaRaw>;